refactored config fields and directory structure
