+++ crates/opennote-bootstrap/Cargo.toml
@@ -9,7 +9,6 @@ license.workspace = true
 opennote-embedder = { workspace = true }
 opennote-models = { workspace = true }
 opennote-data = { workspace = true }
-opennote-tasks-scheduler = { workspace = true }
 
 # External dependencies
 tokio = { workspace = true }
