+++ crates/opennote-models/src/configurations/system.rs
@@ -3,24 +3,13 @@
 //! Modifications to these may incur break changes to the existing database.
 
 use serde::{Deserialize, Serialize};
-use serde_encrypt::shared_key::SharedKey;
 
-use crate::{
-    constants::{
-        DATA_STORAGE_FOLDER_NAME, SQLITE_VECTOR_DATABASE_FILE_EXTENSION, VECTOR_DATABASE_FILENAME,
-        env_vars::{DEFAULT_SQLITE_DATA_FOLDER_NAME_ENV_VAR_NAME, load_environment_variable},
-    },
-    providers::{
-        database::DatabaseProvider, embedder::EmbedderProvider,
-        vector_database::VectorDatabaseProvider,
-    },
+use crate::configurations::fields::{
+    DatabaseConfig, EmbedderConfig, LoggingConfig, VectorDatabaseConfig,
 };
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SystemConfigurations {
-    /// Server related settings. It will be ignored when using in apps
-    pub server: ServerConfig,
-
     /// Logging settings
     pub logging: LoggingConfig,
 
@@ -34,167 +23,9 @@ pub struct SystemConfigurations {
     pub embedder: EmbedderConfig,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EmbedderConfig {
-    /// Provider of the embedding model
-    pub provider: EmbedderProvider,
-
-    /// base url of your local embedder service.
-    pub base_url: String,
-
-    /// Model name of the embedding model
-    pub model: String,
-
-    /// Larger number will make the vectorization faster,
-    /// but try reducing the number to prevent overflowing the API
-    pub vectorization_batch_size: usize,
-
-    /// Dimension of the embedding model
-    pub dimensions: usize,
-
-    /// Usually this is a float
-    pub encoding_format: String,
-
-    /// API key of the model
-    pub api_key: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DatabaseConfig {
-    pub provider: DatabaseProvider,
-    pub connection_url: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VectorDatabaseConfig {
-    pub provider: VectorDatabaseProvider,
-    pub index: String,
-    pub base_url: String,
-    pub api_key: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ServerConfig {
-    pub host: String,
-    pub port: u16,
-    pub workers: usize,
-
-    /// The shared key is set for this server
-    pub shared_key: SharedKey,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LoggingConfig {
-    pub level: LoggingLevel,
-    pub format: LoggingFormat,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum LoggingFormat {
-    Json,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum LoggingLevel {
-    Trace,
-    Debug,
-    Info,
-    Warn,
-    Error,
-}
-
-impl Default for LoggingConfig {
-    fn default() -> Self {
-        Self {
-            level: LoggingLevel::Info,
-            format: LoggingFormat::Json,
-        }
-    }
-}
-
-impl Default for ServerConfig {
-    fn default() -> Self {
-        Self {
-            host: "0.0.0.0".to_string(),
-            port: 8080,
-            workers: 4,
-            shared_key: SharedKey::new_const([0u8; 32]),
-        }
-    }
-}
-
-impl Default for DatabaseConfig {
-    fn default() -> Self {
-        if let Some(config_dir) = dirs::config_dir() {
-            let app_data_folder_name =
-                load_environment_variable(DEFAULT_SQLITE_DATA_FOLDER_NAME_ENV_VAR_NAME);
-
-            // Looks like this but should be an absolute path:
-            // sqlite://./data/database.sqlite?mode=rwc
-            let path_to_sqlite = config_dir
-                .join(app_data_folder_name)
-                .join(DATA_STORAGE_FOLDER_NAME)
-                .join("database.sqlite")
-                .to_string_lossy()
-                .to_string();
-
-            return Self {
-                provider: DatabaseProvider::SQLite,
-                connection_url: format!("sqlite://{}?mode=rwc", path_to_sqlite),
-            };
-        }
-
-        panic!("No config directory was found in this system");
-    }
-}
-
-impl Default for VectorDatabaseConfig {
-    fn default() -> Self {
-        if let Some(config_dir) = dirs::config_dir() {
-            let app_data_folder_name =
-                load_environment_variable(DEFAULT_SQLITE_DATA_FOLDER_NAME_ENV_VAR_NAME);
-
-            // Looks like this but should be an absolute path:
-            // ./data
-            let mut vector_database_path = config_dir
-                .join(app_data_folder_name)
-                .join(DATA_STORAGE_FOLDER_NAME)
-                .join(VECTOR_DATABASE_FILENAME);
-
-            vector_database_path.add_extension(SQLITE_VECTOR_DATABASE_FILE_EXTENSION);
-
-            return Self {
-                provider: VectorDatabaseProvider::SQLiteVector,
-                index: "opennote".to_string(),
-                base_url: vector_database_path.to_string_lossy().to_string(),
-                api_key: "".to_string(),
-            };
-        }
-
-        panic!("No config directory was found in this system");
-    }
-}
-
-impl Default for EmbedderConfig {
-    fn default() -> Self {
-        Self {
-            provider: EmbedderProvider::Native,
-            base_url: "".to_string(),
-            model: "sentence-transformers/all-MiniLM-L6-v2".to_string(),
-            vectorization_batch_size: 100, // How many vectorization tasks at a time
-            dimensions: 384, // sentence-transformers/all-MiniLM-L6-v2 is a 1024 dimensional model
-            encoding_format: "float".to_string(),
-            api_key: "".to_string(),
-        }
-    }
-}
-
 impl Default for SystemConfigurations {
     fn default() -> Self {
         Self {
-            server: ServerConfig::default(),
             logging: LoggingConfig::default(),
             database: DatabaseConfig::default(),
             vector_database: VectorDatabaseConfig::default(),
